requirements.txt
torchvision==0.10.0
matplotlib==3.4.2
torch==1.9.0
Pillow==8.3.2
